add some clean

+++ ft_memchr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_memchr(const void *src, int c, int n)
+void	*ft_memchr(const void *src, int c, size_t n)
 {
 	unsigned char	*s;
